Update API POST /{partnerId}/policies/{policyId}/credential-types-request #1861
