+++ partner/partner-management-service/src/main/java/io/mosip/pms/partner/controller/PartnerServiceController.java
@@ -85,9 +85,6 @@ public class PartnerServiceController {
 	@Value("${mosip.pms.api.id.create.partner.post}")
 	private String postCreatePartnerId;
 
-	@Value("${mosip.pms.api.id.partners.credentialtypes.request.post}")
-	private String postPartnerCredentialTypesRequestId;
-
 	@Value("${mosip.pms.api.id.partner.exists.post}")
 	private String postPartnerExistsId;
 
@@ -181,25 +178,6 @@ public class PartnerServiceController {
 		return new ResponseEntity<>(response, HttpStatus.OK);		
 	}
 
-	@PreAuthorize("hasAnyRole(@authorizedRoles.getPostpartnersbioextractors())")
-	@RequestMapping(value = "/{partnerId}/policies/{policyId}/credential-types-request", method = RequestMethod.POST)
-	@Operation(summary = "Service to submit credential types request",
-			description = "Persists credential type request against an in-progress partner policy mapping request")
-	public ResponseWrapperV2<String> submitCredentialTypesRequest(
-			@PathVariable String partnerId,
-			@PathVariable String policyId,
-			@RequestBody @Valid RequestWrapperV2<CredentialTypeRequestDto> requestWrapper) {
-		Optional<ResponseWrapperV2<String>> validationResponse =
-				requestValidator.validate(postPartnerCredentialTypesRequestId, requestWrapper);
-		if (validationResponse.isPresent()) {
-			return validationResponse.get();
-		}
-		ResponseWrapperV2<String> response = new ResponseWrapperV2<>();
-		response.setResponse(partnerService.submitCredentialTypesRequest(partnerId, policyId, requestWrapper.getRequest()));
-		response.setId(requestWrapper.getId());
-		response.setVersion(requestWrapper.getVersion());
-		return response;
-	}
 
 	/**
 	 * 
@@ -218,14 +196,14 @@ public class PartnerServiceController {
 	}
 	
 	/**
-	 * @deprecated Replaced by POST /partners/{partnerId}/policies/{policyId}/credential-types-request.
+	 * @deprecated Replaced by POST /partner-policy-requests/{requestId}/credential-types-request.
 	 */
 	@Deprecated(since = "release-1.3.0-beta.5")
 	@PreAuthorize("hasAnyRole(@authorizedRoles.getPostpartnerscredentialtypepolicies())")
 	@RequestMapping(value = "/{partnerId}/credentialtype/{credentialType}/policies/{policyName}",method = RequestMethod.POST)
 	@Operation(
 			summary = "Service to map partner and policy to a credential type — deprecated since release-1.3.0-beta.5",
-			description = "Deprecated since release-1.3.0-beta.5. Replaced by POST /partners/{partnerId}/policies/{policyId}/credential-types-request.",
+			description = "Deprecated since release-1.3.0-beta.5. Replaced by POST /partner-policy-requests/{requestId}/credential-types-request.",
 			deprecated = true)
 	public ResponseEntity<ResponseWrapper<String>> mapPolicyToCredentialType(@PathVariable @Valid String partnerId ,@PathVariable @Valid String policyName,
 			@PathVariable @Valid String credentialType){
