+++ partner/partner-management-service/src/main/java/io/mosip/pms/partner/service/PartnerService.java
@@ -146,7 +146,7 @@ public interface PartnerService {
 	 * @param request credential type request payload (includes partnerPolicyRequestId)
 	 * @return success message
 	 */
-	String submitCredentialTypesRequest(String partnerId, String policyId, CredentialTypeRequestDto request);
+	String submitCredentialTypesRequest(CredentialTypeRequestDto request);
 
     /**
      * Function to get biometric extractors of partner and policy
